rpk: Add 'rpk redpanda stop', deprecate 'rpk stop'
